Moved contractnet example to platform demo folder
